Refactor create() out of webpack.js
